# Server/MagicStreamMoviesServer/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for MagicStreamMovies application.",
        "title": "MagicStreamMovies API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/login": {
            "post": {
                "description": "Login user and return tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "User Login Credentials",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "description": "Register a new user with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User Registration Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "description": "Logout user and clear tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Logout user",
                "parameters": [
                    {
                        "description": "User ID JSON {\\",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/user/refresh-token": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Refresh access and refresh tokens",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Refresh user tokens",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.Genre": {
            "type": "object",
            "required": [
                "genre_id",
                "genre_name"
            ],
            "properties": {
                "genre_id": {
                    "type": "integer"
                },
                "genre_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        },
...
        "github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.User": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "password",
                "role"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "expires_at": {
                    "$ref": "#/definitions/go_token.Token"
                },
                "favourite_genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.Genre"
                    }
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "refresh_token": {
                    "type": "string"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "ADMIN",
                        "USER"
                    ]
                },
                "token": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.UserLogin": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.UserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "favourite_genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sirolad_MagicStreamMovies_Server_MagicStreamMovies_Server_internal_models.Genre"
                    }
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "go_token.Token": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14,
                15,
                16,
                17,
                18,
                19,
                20,
                21,
                22,
                23,
                24,
                25,
                26,
                27,
                28,
                29,
                30,
                31,
                32,
                33,
                34,
                35,
                36,
                37,
                38,
                39,
                40,
                41,
                42,
                43,
                44,
                45,
                46,
                47,
                48,
                49,
                50,
                51,
                52,
                53,
                54,
                55,
                56,
                57,
                58,
                59,
                60,
                61,
                62,
                63,
                64,
                65,
                66,
                67,
                68,
                69,
                70,
                71,
                72,
                73,
                74,
                75,
                76,
                77,
                78,
                79,
                80,
                81,
                82,
                83,
                84,
                85,
                86,
                87,
                88,
                89
            ],
            "x-enum-comments": {
                "ADD": "+",
                "ADD_ASSIGN": "+=",
                "AND": "\u0026",
                "AND_ASSIGN": "\u0026=",
                "AND_NOT": "\u0026^",
                "AND_NOT_ASSIGN": "\u0026^=",
                "ARROW": "\u003c-",
                "ASSIGN": "=",
                "CHAR": "'a'",
                "COLON": ":",
                "COMMA": ",",
                "DEC": "--",
                "DEFINE": ":=",
                "ELLIPSIS": "...",
                "EQL": "==",
                "FLOAT": "123.45",
                "GEQ": "\u003e=",
                "GTR": "\u003e",
                "IDENT": "main",
                "IMAG": "123.45i",
                "INC": "++",
                "INT": "12345",
                "LAND": "\u0026\u0026",
                "LBRACE": "{",
                "LBRACK": "[",
                "LEQ": "\u003c=",
                "LOR": "||",
                "LPAREN": "(",
                "LSS": "\u003c",
                "MUL": "*",
                "MUL_ASSIGN": "*=",
                "NEQ": "!=",
                "NOT": "!",
                "OR": "|",
                "OR_ASSIGN": "|=",
                "PERIOD": ".",
                "QUO": "/",
                "QUO_ASSIGN": "/=",
                "RBRACE": "}",
                "RBRACK": "]",
                "REM": "%",
                "REM_ASSIGN": "%=",
                "RPAREN": ")",
                "SEMICOLON": ";",
                "SHL": "\u003c\u003c",
                "SHL_ASSIGN": "\u003c\u003c=",
                "SHR": "\u003e\u003e",
                "SHR_ASSIGN": "\u003e\u003e=",
                "STRING": "\"abc\"",
                "SUB": "-",
                "SUB_ASSIGN": "-=",
                "XOR": "^",
                "XOR_ASSIGN": "^="
            },
            "x-enum-descriptions": [
                "",
                "",
                "",
                "",
                "main",
                "12345",
                "123.45",
                "123.45i",
                "'a'",
                "\"abc\"",
                "",
                "",
                "+",
                "-",
                "*",
                "/",
                "%",
                "\u0026",
                "|",
                "^",
                "\u003c\u003c",
                "\u003e\u003e",
                "\u0026^",
                "+=",
                "-=",
                "*=",
                "/=",
                "%=",
                "\u0026=",
                "|=",
                "^=",
                "\u003c\u003c=",
                "\u003e\u003e=",
                "\u0026^=",
                "\u0026\u0026",
                "||",
                "\u003c-",
                "++",
                "--",
                "==",
                "\u003c",
                "\u003e",
                "=",
                "!",
                "!=",
                "\u003c=",
                "\u003e=",
                ":=",
                "...",
                "(",
                "[",
                "{",
                ",",
                ".",
                ")",
                "]",
                "}",
                ";",
                ":",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                ""
            ],
            "x-enum-varnames": [
                "ILLEGAL",
                "EOF",
                "COMMENT",
                "literal_beg",
                "IDENT",
                "INT",
                "FLOAT",
                "IMAG",
                "CHAR",
                "STRING",
                "literal_end",
                "operator_beg",
                "ADD",
                "SUB",
                "MUL",
                "QUO",
                "REM",
                "AND",
                "OR",
                "XOR",
                "SHL",
                "SHR",
                "AND_NOT",
                "ADD_ASSIGN",
                "SUB_ASSIGN",
                "MUL_ASSIGN",
                "QUO_ASSIGN",
                "REM_ASSIGN",
                "AND_ASSIGN",
                "OR_ASSIGN",
                "XOR_ASSIGN",
                "SHL_ASSIGN",
                "SHR_ASSIGN",
                "AND_NOT_ASSIGN",
                "LAND",
                "LOR",
                "ARROW",
                "INC",
                "DEC",
                "EQL",
                "LSS",
                "GTR",
                "ASSIGN",
                "NOT",
                "NEQ",
                "LEQ",
                "GEQ",
                "DEFINE",
                "ELLIPSIS",
                "LPAREN",
                "LBRACK",
                "LBRACE",
                "COMMA",
                "PERIOD",
                "RPAREN",
                "RBRACK",
                "RBRACE",
                "SEMICOLON",
                "COLON",
                "operator_end",
                "keyword_beg",
                "BREAK",
                "CASE",
                "CHAN",
                "CONST",
                "CONTINUE",
                "DEFAULT",
                "DEFER",
                "ELSE",
                "FALLTHROUGH",
                "FOR",
                "FUNC",
                "GO",
                "GOTO",
                "IF",
                "IMPORT",
                "INTERFACE",
                "MAP",
                "PACKAGE",
                "RANGE",
                "RETURN",
                "SELECT",
                "STRUCT",
                "SWITCH",
                "TYPE",
                "VAR",
                "keyword_end",
                "additional_beg",
                "TILDE",
                "additional_end"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your Bearer token (e.g., \"eyJhbGci...\"). The \"Bearer \" prefix is optional.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}