# Server/MagicStreamMoviesServer/internal/models/user_model.go
package models

import (
	"go/token"
	"time"

	"go.mongodb.org/mongo-driver/v2/bson"
)

type User struct {
	ID             bson.ObjectID `json:"id,omitempty" bson:"_id,omitempty"`
	UserID         string        `json:"user_id" bson:"user_id"`
	FirstName      string        `json:"first_name" bson:"first_name" validate:"required,min=2,max=100"`
	LastName       string        `json:"last_name" bson:"last_name" validate:"required,min=2,max=100"`
	Email          string        `json:"email" bson:"email" validate:"required,email"`
	Password       string        `json:"password" bson:"password" validate:"required,min=6"`
	Role           string        `json:"role" bson:"role" validate:"required,oneof=ADMIN USER"`
	CreatedAt      time.Time     `json:"created_at" bson:"created_at"`
	UpdatedAt      time.Time     `json:"updated_at" bson:"updated_at"`
	Token          string        `json:"token" bson:"token"`
	RefreshToken   string        `json:"refresh_token" bson:"refresh_token"`
	ExpiresAt      *token.Token  `json:"expires_at" bson:"expires_at"`
	FavoriteGenres []Genre       `json:"favourite_genres" bson:"favourite_genres" validate:"dive"`
}

type UserLogin struct {
	Email    string `json:"email" validate:"required,email"`
	Password string `json:"password" validate:"required,min=6"`
}

type UserResponse struct {
	UserID         string  `json:"user_id"`
	FirstName      string  `json:"first_name"`
	LastName       string  `json:"last_name"`
	Email          string  `json:"email"`
	Role           string  `json:"role"`
	Token          string  `json:"token"`
	RefreshToken   string  `json:"refresh_token"`
	FavoriteGenres []Genre `json:"favourite_genres"`
}